pyproject.toml
[project]
name = "spatiotemporal_lab"
version = "0.1.0"
description = "Spatiotemporal dataset and benchmarks for traffic forecasting, environmental modeling, and multi-objective optimization."
readme = "README.md"
requires-python = ">=3.13"
# Runtime dependencies that are safe to install in CI (CPU-only, not huge CUDA stacks)
dependencies = [
    "numpy>=1.26",
    "pydantic>=2.8",
    "typing-extensions>=4.7",
    "loguru>=0.7",
    "rich>=13.0",
    "dvc>=3.0",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    # add other core, CPU-safe libs here
]

# Optional feature sets (extras) not installed by default
[project.optional-dependencies]
gpu = [
    # CUDA / GPU-specific deps. NEVER installed in CI.
    # Adapt versions / wheels to your stack as needed.
    "torch==2.9.1+cu130 ; sys_platform == 'linux'",
    "torchvision==0.24.1+cu130 ; sys_platform == 'linux'",
    "xformers>=0.0.28 ; sys_platform == 'linux'",
]


notebooks = [
    "ipykernel",
    "jupyter>=1.0",
    "matplotlib>=3.8",
]
mlops = [
    "dvc-s3>=3.0",
    "mlflow>=3.0",
]
bigmodels = ["transformers", "datasets"]

api = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "mlflow>=3.0",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "tenacity>=8.2",
  "prometheus-client>=0.20",
  "orjson>=3.10",
]

# uv dependency groups for tools used during development / CI
[dependency-groups]
dev = [
    # testing
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-xdist>=3.6",

    # typing
    "mypy>=1.10",

    # linting / formatting / spelling
    "ruff>=0.8.0",
    "codespell>=2.3",

    # security / deps
    "pip-audit>=2.5.5",
    "pre-commit>=3.5",

    # notebooks
    "nbstripout>=0.8",
    "nbdime>=4.0",

    # dev
    "cookiecutter>=2.6",
    "dvc>=3.0",
]

docs-mkdocs = [
  "mkdocs>=1.6",
  # pick ONE theme if you use it
  "mkdocs-material>=9.5",
  # optional, only if you use them:
  # "mkdocstrings[python]>=0.25",
  # "mkdocs-gen-files>=0.5",
  # "mkdocs-literate-nav>=0.6",
  # "mkdocs-awesome-pages-plugin>=2.9",
]

docs-sphinx = [
  "sphinx>=7.0",
  # pick ONE theme if you use it
  "sphinx-rtd-theme>=2.0",
  # optional, only if you use it:
  # "myst-parser>=3.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"



[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux'" },
  { index = "pypi", marker = "sys_platform != 'linux'" },
]
torchvision = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux'" },
  { index = "pypi", marker = "sys_platform != 'linux'" },
]
# Ruff configuration (shared across repos)
[tool.ruff]
# basic settings
line-length = 88
# CRITICAL: Change this to match your project's lowest supported version
# e.g., "py310", "py311", "py312"
target-version = "py313" 
extend-include = ["*.ipynb"]

[tool.ruff.lint]
# Enable the rules
# E = pycodestyle (standard style)
# F = Pyflakes (potential bugs)
# I = isort (import sorting)
select = ["E", "F", "I"] 
# Ignore E501 (line too long) because the formatter handles that automatically
ignore = ["E501"] 

[tool.ruff.format]
# Formatter settings (Exact match for Black)
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Pytest defaults + markers for CI to skip GPU/slow tests
[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
    "gpu: tests that require a GPU",
    "slow: slow-running tests",
]

# Optional: mypy defaults
[tool.mypy]
python_version = "3.13"
strict = true
mypy_path = ["src"]
ignore_missing_imports = true